proxy auth seems to be working
