+++ pkg/services/k8s/apiserver/storage.go
@@ -520,7 +520,6 @@ func (s *entityStorage) getSignedInUser(ctx context.Context, obj runtime.Object)
 	var err error
 	userQuery := userpkg.GetSignedInUserQuery{}
 
-	s.log.Warn("user", "user", user)
 	if user.GetName() == authnz.ApiServerUser || user.GetName() == authnz.ApiServerAnonymous {
 		userQuery.OrgID = 1
 		userQuery.UserID = 1
